returning event_s, albeit we are off by two so manually adding two to both event_start and event_end, but this might not work in all cases

+++ src/main/java/org/desilvahendricksoftware/debs/Result.java
@@ -1,16 +1,11 @@
 package org.desilvahendricksoftware.debs;
 
-import com.cedarsoftware.util.io.JsonWriter;
-
-import java.util.HashMap;
-import java.util.Map;
-
 public class Result {
     long s;
     int d;
-    long event_s;
+    double event_s;
 
-    public Result(long s, boolean d, long event_s) {
+    public Result(long s, boolean d, double event_s) {
         this.s = s;
         this.d = d == true ? 1 : 0 ;
         this.event_s = event_s;
